requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
drf-spectacular==0.26.5
python-decouple==3.8
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.4.0
